+++ src/main/java/com/stock/invest/enums/dto/StockDailyBarCandleDto.java
@@ -1,13 +1,15 @@
 package com.stock.invest.enums.dto;
 
+import java.math.BigDecimal;
+
 public record StockDailyBarCandleDto(
     String date,           // 格式: yyyy-MM-dd
-    Double open,           // 开盘价
-    Double high,           // 最高价
-    Double low,            // 最低价
-    Double close,          // 收盘价
-    Double changePercent,  // 涨跌幅（%）
-    Double afterHours,     // 盘后价
-    Double afterHoursChangePercent,  // 盘后涨跌幅（%）
+    BigDecimal open,       // 开盘价
+    BigDecimal high,       // 最高价
+    BigDecimal low,        // 最低价
+    BigDecimal close,      // 收盘价
+    BigDecimal changePercent,  // 涨跌幅（%）
+    BigDecimal afterHours,     // 盘后价
+    BigDecimal afterHoursChangePercent,  // 盘后涨跌幅（%）
     Long volume            // 成交量
 ) {}
